+++ src/components/Title/Title.jsx
@@ -1,0 +1,3 @@
+export const Title = ({ title, styles }) => {
+	return <span className={styles}>{title}</span>;
+};
