+++ src/constants.js
@@ -1,19 +1,34 @@
 export const ANIMATION = {
-  DURATION: 10,
+	DURATION: 10,
 };
 
 export const months = {
-  January: 31,
-  February: 28,
-  March: 31,
-  April: 30,
-  May: 31,
-  June: 30,
-  July: 31,
-  August: 31,
-  September: 30,
-  October: 31,
-  November: 30,
-  December: 31,
-  DURATION: 50,
+	January: 31,
+	February: 28,
+	March: 31,
+	April: 30,
+	May: 31,
+	June: 30,
+	July: 31,
+	August: 31,
+	September: 30,
+	October: 31,
+	November: 30,
+	December: 31,
+	DURATION: 50,
 };
+
+export const monthsName = [
+	'January',
+	'February',
+	'March',
+	'April',
+	'May',
+	'June',
+	'July',
+	'August',
+	'September',
+	'October',
+	'November',
+	'December',
+];
